requirements.txt
# Core Framework
streamlit>=1.28.0

# Environment
python-dotenv>=1.0.0

# LangChain & LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0

# Hugging Face (for LLM inference)
huggingface_hub>=0.20.0
transformers>=4.30.0

# Vector Database (FAISS)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Search Providers
tavily-python>=0.3.0
google-search-results>=2.4.2
duckduckgo-search>=6.0.0

# LangSmith (observability)
langsmith>=0.1.0

# Export Features
fpdf2>=2.7.0
markdown>=3.5.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0
